# docs/swagger.yaml
basePath: /api/v1
definitions:
  user.CreateUserDTO:
    properties:
      email:
        example: aman@gmail.com
        type: string
      name:
        example: Aman
        type: string
    required:
    - email
    - name
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:3010
info:
  contact:
    email: aman@example.com
    name: Aman
  description: Production ready Fiber backend
  termsOfService: https://example.com/terms
  title: Fiber Production API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with search
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Create user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - basicAuth: []
      summary: Create a new user
      tags:
      - Users
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"